[Wen Hao] - task 27.
(cherry picked from commit 0d9e12f)

+++ 04-taxi/src/main/java/com/github/wenhao/AdditionalFeeTaxiCharge.java
@@ -13,7 +13,7 @@ public class AdditionalFeeTaxiCharge implements TaxiCharge
 
     public AdditionalFeeTaxiCharge(final double baseMeter, final BigDecimal pricePerMile)
     {
-        this(baseMeter, Integer.MAX_VALUE, pricePerMile);
+        this(baseMeter, Double.MAX_VALUE, pricePerMile);
     }
 
     public AdditionalFeeTaxiCharge(final double minMeter, final double maxMeter, final BigDecimal pricePerMile)
